+++ Buffer/CborInputCyclicBuffer.cs
@@ -275,9 +275,9 @@ namespace Naphaso.Cbor.Buffer
                 return r;
             }
 
-            return ((ulong)this.GetByte() << 56) | ((ulong)this.GetByte() << 48) | ((ulong)this.GetByte() << 40)
-                   | ((ulong)this.GetByte() << 32) | ((ulong)this.GetByte() << 24) | ((ulong)this.GetByte() << 16)
-                   | ((ulong)this.GetByte() << 8) | ((ulong)this.GetByte());
+            return (ulong)((ulong)this.GetByte() << 56) | (ulong)((ulong)this.GetByte() << 48) | (ulong)((ulong)this.GetByte() << 40)
+                   | (ulong)((ulong)this.GetByte() << 32) | (ulong)((ulong)this.GetByte() << 24) | (ulong)((ulong)this.GetByte() << 16)
+                   | (ulong)((ulong)this.GetByte() << 8) | (ulong)((ulong)this.GetByte());
         }
 
         /// <summary>
